+++ src/Certes.Cli/Commands/CertificatePfxCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.CommandLine;
 using System.Globalization;
+using System.Text;
 using System.Threading.Tasks;
 using Certes.Cli.Settings;
 using NLog;
@@ -13,6 +14,7 @@ namespace Certes.Cli.Commands
         private const string PasswordParam = "password";
         private const string PrivateKeyOption = "private-key";
         private const string OutOption = "out";
+        private const string IssuerOption = "issuer";
         private static readonly ILogger logger = LogManager.GetLogger(nameof(CertificatePfxCommand));
 
         private readonly IEnvironmentVariables environment;
@@ -36,6 +38,7 @@ namespace Certes.Cli.Commands
                 .DefineKeyOption()
                 .DefineOption(OutOption, help: Strings.HelpCertificateOut)
                 .DefineOption(PrivateKeyOption, help: Strings.HelpPrivateKey)
+                .DefineOption(IssuerOption, help: Strings.HelpCertificateIssuer)
                 .DefineUriParameter(OrderIdParam, help: Strings.HelpOrderId)
                 .DefineParameter(PasswordParam, help: Strings.HelpPfxPassword);
 
@@ -46,11 +49,20 @@ namespace Certes.Cli.Commands
         {
             var keyPath = syntax.GetParameter<string>(PrivateKeyOption, true);
             var pwd = syntax.GetParameter<string>(PasswordParam, true);
+            var issuer = syntax.GetParameter<string>(IssuerOption);
             var (location, cert) = await DownloadCertificate(syntax);
 
             var pfxName = string.Format(CultureInfo.InvariantCulture, "[certes] {0:yyyyMMddhhmmss}", DateTime.UtcNow);
             var privKey = await syntax.ReadKey(PrivateKeyOption, "CERTES_CERT_KEY", File, environment, true);
-            var pfx = cert.ToPfx(privKey).Build(pfxName, pwd);
+
+            var pfxBuilder = cert.ToPfx(privKey);
+            if (!string.IsNullOrWhiteSpace(issuer))
+            {
+                var issuerPem = await File.ReadAllText(issuer);
+                pfxBuilder.AddIssuers(Encoding.UTF8.GetBytes(issuerPem));
+            }
+
+            var pfx = pfxBuilder.Build(pfxName, pwd);
             
             var outPath = syntax.GetOption<string>(OutOption);
             if (string.IsNullOrWhiteSpace(outPath))